Add some doc comments and tests to merkle.rs
I definitely had some incorrect assumptions about this data structure
which made it more difficult to learn. So, I'm documenting how it
works and adding some tests.

The simple_merkle test is currently failing because the `set` method
doesn't allow setting an index larger than the largest currently
set leaf's index.

There is some debate as to whether or not this is the correct
behavior. To run the test, use:

```
$> cargo test -- --include-ignored
```