Change the implementation to a flat vector of bytes.
The idea behind this change was to increase Layer 2 cache hits by
storing all of the memory related to the tee in a contiguous block.

Unfortunately, as of this commit, this produces a considerable
performance degredation over the implementaiton with a
`Vec<Vec<Byte32>>` representing the layers in the tree.

The implementation is working correctly. But, it's slow.

I'm going to run a profiler on it now and see if there are some
obvious bottlenecks.